feat: add toolbar and download search functionality to offline/no internet activity
