+++ app/src/main/java/com/github/libretube/ui/activities/AbstractPlayerHostActivity.kt
@@ -1,5 +1,9 @@
 package com.github.libretube.ui.activities
 
+import android.content.Intent
+import android.view.MenuItem
+import com.github.libretube.R
+import com.github.libretube.helpers.IntentHelper
 import com.github.libretube.ui.base.BaseActivity
 import com.github.libretube.ui.fragments.AudioPlayerFragment
 import com.github.libretube.ui.fragments.PlayerFragment
@@ -28,5 +32,42 @@ abstract class AbstractPlayerHostActivity: BaseActivity() {
             .firstOrNull()
             ?.let(action)
             ?: false
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        return when (item.itemId) {
+            R.id.action_settings -> {
+                val settingsIntent = Intent(this, SettingsActivity::class.java)
+                startActivity(settingsIntent)
+                true
+            }
+
+            R.id.action_about -> {
+                val aboutIntent = Intent(this, AboutActivity::class.java)
+                startActivity(aboutIntent)
+                true
+            }
+
+            R.id.action_help -> {
+                val helpIntent = Intent(this, HelpActivity::class.java)
+                startActivity(helpIntent)
+                true
+            }
+
+            R.id.action_donate -> {
+                IntentHelper.openLinkFromHref(
+                    this,
+                    supportFragmentManager,
+                    AboutActivity.DONATE_URL,
+                    forceDefaultOpen = true
+                )
+                true
+            }
+
+            else -> super.onOptionsItemSelected(item)
+        }
     }
 }
