+++ app/src/main/java/com/github/libretube/ui/fragments/NoInternetFragment.kt
@@ -1,6 +1,5 @@
 package com.github.libretube.ui.fragments
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
@@ -9,7 +8,6 @@ import com.github.libretube.R
 import com.github.libretube.databinding.FragmentNointernetBinding
 import com.github.libretube.helpers.NavigationHelper
 import com.github.libretube.helpers.NetworkHelper
-import com.github.libretube.ui.activities.SettingsActivity
 import com.google.android.material.snackbar.Snackbar
 
 class NoInternetFragment: Fragment(R.layout.fragment_nointernet) {
@@ -24,10 +22,6 @@ class NoInternetFragment: Fragment(R.layout.fragment_nointernet) {
             } else {
                 Snackbar.make(binding.root, R.string.turnInternetOn, Snackbar.LENGTH_LONG).show()
             }
-        }
-        binding.noInternetSettingsImageView.setOnClickListener {
-            val intent = Intent(requireContext(), SettingsActivity::class.java)
-            startActivity(intent)
         }
 
         binding.downloads.setOnClickListener {
